pref: 优化平台
