+++ apps/assets/const/__init__.py
@@ -1,0 +1,3 @@
+from .protocol import *
+from .category import *
+from .types import *
